feat: consolidate backend domain and persistence baseline

+++ src/main/java/com/jefiro/app247/infra/exception/PriceChangedException.java
@@ -28,7 +28,7 @@ public class PriceChangedException extends RuntimeException {
             boolean emPromocao,
             String promocaoId,
             String promocaoNome,
-            int quantidade,
+            BigDecimal quantidade,
             boolean aumentou
     ) {
     }
